// src/app/core/models/codeHelper.ts
/**************************************************************************
*  Revision History:
*
**************************************************************************/



export interface CodeHelper {
    codeType: string;
    codeValue: string;
    codeSubType?: string;
    codeDesc: string;
    codeOrder: number;
    codeFib?: string;
    codeFid?: Date;
    codeLmb?: string;
    codeLmd?: Date;
    toolTip?: string;
}